move find_executable and remove unused script
