+++ tests/test_execution.py
@@ -4,7 +4,7 @@ import subprocess
 import unittest
 from importlib.resources import files
 from app.models.execution import Execution
-from app.models.find_executable import get_pea_exec
+from app.utils.find_executable import get_pea_exec
 
 class TestExecution(unittest.TestCase):
     def test_load_config(self):
